+++ test-e2e/cypress/support/commands.js
@@ -1,6 +1,1 @@
-import { config } from "@mocks-server/cypress-commands";
-
-config({
-  adminApiPath: Cypress.env("MOCKS_ADMIN_API_PATH"),
-  baseUrl: Cypress.env("MOCKS_URL"),
-});
+import "@mocks-server/cypress-commands";