+++ test/commands.spec.js
@@ -1,29 +1,73 @@
 const sinon = require("sinon");
-const adminApiClient = require("@mocks-server/admin-api-client");
+const apiClient = require("@mocks-server/admin-api-client");
 
-const { setBehavior, setDelay, setSettings, config } = require("../src/commands");
+const CypressMock = require("./Cypress.mock");
+
+const commands = require("../src/commands");
 
 describe("commands", () => {
   let sandbox;
+  let cypressMock;
+  let setMock, setBehavior, setDelay, setSettings, useRouteVariant, restoreRoutesVariants, config;
 
   beforeEach(() => {
     sandbox = sinon.createSandbox();
-    sandbox.stub(adminApiClient.settings, "update");
-    sandbox.stub(adminApiClient, "config");
+    sandbox.stub(apiClient, "updateSettings");
+    sandbox.stub(apiClient, "useRouteVariant");
+    sandbox.stub(apiClient, "restoreRoutesVariants");
+    sandbox.stub(apiClient, "config");
+    cypressMock = new CypressMock();
+    ({
+      setMock,
+      setBehavior,
+      setDelay,
+      setSettings,
+      useRouteVariant,
+      restoreRoutesVariants,
+      config,
+    } = commands(cypressMock.stubs));
   });
 
   afterEach(() => {
     sandbox.restore();
   });
 
-  describe("setBehavior command", () => {
-    it("should call to update behavior", () => {
-      setBehavior("foo-behavior");
+  describe("when initializing", () => {
+    it("should call to set adminApiPath config if env var is defined", () => {
+      cypressMock.stubs.env.withArgs("MOCKS_SERVER_ADMIN_API_PATH").returns("foo");
+      commands(cypressMock.stubs);
       expect(
-        adminApiClient.settings.update.calledWith({
-          behavior: "foo-behavior",
+        apiClient.config.calledWith({
+          adminApiPath: "foo",
         })
       ).toBe(true);
+    });
+
+    it("should call to set baseUrl config if env var is defined", () => {
+      cypressMock.stubs.env.withArgs("MOCKS_SERVER_BASE_URL").returns("foo");
+      commands(cypressMock.stubs);
+      expect(
+        apiClient.config.calledWith({
+          baseUrl: "foo",
+        })
+      ).toBe(true);
+    });
+  });
+
+  describe("setMock command", () => {
+    it("should call to update delay", () => {
+      setMock("foo");
+      expect(
+        apiClient.updateSettings.calledWith({
+          mock: "foo",
+        })
+      ).toBe(true);
+    });
+
+    it("should do nothing if plugin is disabled", () => {
+      cypressMock.stubs.env.returns("false");
+      setMock("foo");
+      expect(apiClient.updateSettings.callCount).toEqual(0);
     });
   });
 
@@ -31,17 +75,72 @@ describe("commands", () => {
     it("should call to update delay", () => {
       setDelay(3000);
       expect(
-        adminApiClient.settings.update.calledWith({
+        apiClient.updateSettings.calledWith({
           delay: 3000,
         })
       ).toBe(true);
+    });
+
+    it("should do nothing if plugin is disabled", () => {
+      cypressMock.stubs.env.returns("false");
+      setDelay("foo");
+      expect(apiClient.updateSettings.callCount).toEqual(0);
     });
   });
 
   describe("setSettings command", () => {
     it("should call to update delay", () => {
       setSettings("foo");
-      expect(adminApiClient.settings.update.calledWith("foo")).toBe(true);
+      expect(apiClient.updateSettings.calledWith("foo")).toBe(true);
+    });
+
+    it("should do nothing if plugin is disabled", () => {
+      cypressMock.stubs.env.returns("0");
+      setSettings("foo");
+      expect(apiClient.updateSettings.callCount).toEqual(0);
+    });
+  });
+
+  describe("useRouteVariant command", () => {
+    it("should call to useRoute variant", () => {
+      useRouteVariant("foo");
+      expect(apiClient.useRouteVariant.calledWith("foo")).toBe(true);
+    });
+
+    it("should do nothing if plugin is disabled", () => {
+      cypressMock.stubs.env.returns(0);
+      useRouteVariant("foo");
+      expect(apiClient.useRouteVariant.callCount).toEqual(0);
+    });
+  });
+
+  describe("restoreRoutesVariants command", () => {
+    it("should call to useRoute variant", () => {
+      restoreRoutesVariants();
+      expect(apiClient.restoreRoutesVariants.callCount).toEqual(1);
+    });
+
+    it("should do nothing if plugin is disabled", () => {
+      cypressMock.stubs.env.returns(false);
+      restoreRoutesVariants();
+      expect(apiClient.restoreRoutesVariants.callCount).toEqual(0);
+    });
+  });
+
+  describe("setBehavior command", () => {
+    it("should call to update behavior", () => {
+      setBehavior("foo-behavior");
+      expect(
+        apiClient.updateSettings.calledWith({
+          behavior: "foo-behavior",
+        })
+      ).toBe(true);
+    });
+
+    it("should do nothing if plugin is disabled", () => {
+      cypressMock.stubs.env.returns(false);
+      setBehavior("foo-behavior");
+      expect(apiClient.updateSettings.callCount).toEqual(0);
     });
   });
 
@@ -51,7 +150,7 @@ describe("commands", () => {
         baseUrl: "foo",
       });
       expect(
-        adminApiClient.config.calledWith({
+        apiClient.config.calledWith({
           baseUrl: "foo",
         })
       ).toBe(true);
@@ -62,8 +161,8 @@ describe("commands", () => {
         adminApiPath: "foo",
       });
       expect(
-        adminApiClient.config.calledWith({
-          apiPath: "foo",
+        apiClient.config.calledWith({
+          adminApiPath: "foo",
         })
       ).toBe(true);
     });