+++ CHANGELOG.md
@@ -9,6 +9,52 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 ### Fixed
 ### Removed
+### Breaking changes
+
+## [2.0.0] - 2021-02-17
+
+### Added
+- feat: Add `mocksConfig`, `mocksSetMock`, `mocksUseRouteVariant` and `mocksRestoreRoutesVariants` commands
+- feat: Add support for environment variables `MOCKS_SERVER_ENABLED`, `MOCKS_SERVER_BASE_URL` and `MOCKS_SERVER_ADMIN_API_PATH` (#3)
+
+### Changed
+- feat: Rename `mocksServerSetBehavior` command to `mocksSetBehavior`
+- feat: Rename `mocksServerSetDelay` command to `mocksSetDelay`
+- feat: Rename `mocksServerSetSettings` command to `mocksSetSettings`
+- chore(deps): Update dependencies to mocks-server v2 compatible versions
+- chore(deps): Update Cypress to v6 in tests
+
+### Removed
+- feat: Remove `config` method, `cy.mocksConfig` command should be used instead
+- test(e2e): Remove e2e tests using data-provider, as it does not concern to this package
+
+### Breaking changes
+- Rename `mocksServerSetBehavior` command to `mocksSetBehavior`
+- Rename `mocksServerSetDelay` command to `mocksSetDelay`
+- Rename `mocksServerSetSettings` command to `mocksSetSettings`
+- Remove `config` method, `cy.mocksConfig` command should be used instead
+
+## [2.0.0-beta.1] - 2021-02-16
+
+### Added
+- feat: Add `mocksConfig`, `mocksSetMock`, `mocksUseRouteVariant` and `mocksRestoreRoutesVariants` commands
+- feat: Add support for environment variables `MOCKS_SERVER_ENABLED`, `MOCKS_SERVER_BASE_URL` and `MOCKS_SERVER_ADMIN_API_PATH` (#3)
+
+### Changed
+- feat: Rename `mocksServerSetBehavior` command to `mocksSetBehavior`
+- feat: Rename `mocksServerSetDelay` command to `mocksSetDelay`
+- feat: Rename `mocksServerSetSettings` command to `mocksSetSettings`
+- chore(deps): Update dependencies to mocks-server v2 compatible versions
+
+### Removed
+- feat: Remove `config` method, `cy.mocksConfig` command should be used instead
+- test(e2e): Remove e2e tests using data-provider, as it does not concern to this package
+
+### Breaking changes
+- Rename `mocksServerSetBehavior` command to `mocksSetBehavior`
+- Rename `mocksServerSetDelay` command to `mocksSetDelay`
+- Rename `mocksServerSetSettings` command to `mocksSetSettings`
+- Remove `config` method, `cy.mocksConfig` command should be used instead
 
 ## [1.0.8] - 2020-12-21